refactor: 모임 대시보드 연속달성·주의 배지 컴포넌트 공용화
