+++ src/app/gatherings/[id]/dashboard/_components/MemberRankingSection/RankingItem/index.tsx
@@ -3,11 +3,11 @@
 import Image from 'next/image';
 
 import { Button } from '@/components/ui/Button';
-import { HandIcon, StateIcon } from '@/components/ui/Icon';
+import { HandIcon } from '@/components/ui/Icon';
 import { ProgressBar } from '@/components/ui/Progress';
-import { Tag } from '@/components/ui/Tag';
 import { useFallbackImage } from '@/hooks/useFallbackImage';
 
+import { MemberBadge } from '../../MemberBadge';
 import { RankBadge } from '../RankBadge';
 
 import type { AchievementRankingItem } from '@/api/achievements/types';
@@ -39,26 +39,8 @@ export function RankingItem({ item, isMe, streakDays }: RankingItemProps) {
           <div className='flex items-center justify-between'>
             <div className='flex items-center gap-1 md:gap-2'>
               <span className='text-small-02-m md:text-body-01-m text-gray-900'>{item.nickname}</span>
-              {isWarning && (
-                <Tag
-                  variant='deadline'
-                  state='warning'
-                  className='md:text-small-02-m h-[17px] w-[40px] px-1.5 py-0 text-[10px] md:h-[27px] md:w-[66px] md:px-3 md:py-1'
-                >
-                  <StateIcon variant='warning' className='h-3 w-3 md:h-4 md:w-4' />
-                  주의
-                </Tag>
-              )}
-              {!isWarning && streakDays > 0 && (
-                <Tag
-                  variant='deadline'
-                  state='goal'
-                  className='md:text-small-02-m gap-0.3 flex h-[17px] w-10 px-1.5 py-0 text-[10px] md:h-[27px] md:w-[66px] md:gap-1 md:px-3 md:py-1'
-                >
-                  <StateIcon variant='active' className='h-3 w-3 md:h-4 md:w-4' />
-                  {streakDays}일
-                </Tag>
-              )}
+              {isWarning && <MemberBadge type='warning' label='주의' />}
+              {!isWarning && streakDays > 0 && <MemberBadge type='streak' label={`${streakDays}일`} />}
             </div>
             <div className='flex shrink-0 items-center gap-1'>
               <span className='text-small-02-r md:text-body-01-r text-gray-900'>달성률</span>
